start refactoring reduction test: factor out common parts
